Refactor :: move package location
